[Silver I] Title: 쉬운 계단 수, Time: 124 ms, Memory: 14240 KB -BaekjoonHub

+++ 백준/Silver/10844. 쉬운 계단 수/쉬운 계단 수.java
@@ -7,7 +7,7 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         int n = Integer.parseInt(br.readLine());
-        long[][] dp = new long[n + 1][10];
+        int[][] dp = new int[n + 1][10];
         long cnt = 9;
 
         for (int i = 1; i < 10; i++) {
